+++ firestore-stripe-subscriptions/functions/src/index.ts
@@ -27,7 +27,7 @@ const stripe = new Stripe(config.stripeSecretKey, {
   // https://stripe.com/docs/building-plugins#setappinfo
   appInfo: {
     name: 'Firebase firestore-stripe-subscriptions',
-    version: '0.1.2',
+    version: '0.1.3',
   },
 });
 
@@ -91,6 +91,7 @@ exports.createCheckoutSession = functions.firestore
       cancel_url,
       quantity = 1,
       payment_method_types = ['card'],
+      metadata = {},
     } = snap.data();
     try {
       logs.creatingCheckoutSession(context.params.id);
@@ -113,6 +114,10 @@ exports.createCheckoutSession = functions.firestore
             },
           ],
           mode: 'subscription',
+          subscription_data: {
+            trial_from_plan: true,
+            metadata,
+          },
           success_url,
           cancel_url,
         },
@@ -239,6 +244,7 @@ const manageSubscriptionStatusChange = async (
     .doc(subscription.id);
   // Update with new Subscription status
   const subscriptionData: Subscription = {
+    metadata: subscription.metadata,
     role,
     status: subscription.status,
     stripeLink: `https://dashboard.stripe.com${
@@ -252,9 +258,27 @@ const manageSubscriptionStatusChange = async (
       .doc(price.id),
     quantity: subscription.quantity,
     cancel_at_period_end: subscription.cancel_at_period_end,
+    cancel_at: subscription.cancel_at
+      ? admin.firestore.Timestamp.fromMillis(subscription.cancel_at * 1000)
+      : null,
+    canceled_at: subscription.canceled_at
+      ? admin.firestore.Timestamp.fromMillis(subscription.canceled_at * 1000)
+      : null,
+    current_period_start: admin.firestore.Timestamp.fromMillis(
+      subscription.current_period_start * 1000
+    ),
+    current_period_end: admin.firestore.Timestamp.fromMillis(
+      subscription.current_period_end * 1000
+    ),
     created: admin.firestore.Timestamp.fromMillis(subscription.created * 1000),
     ended_at: subscription.ended_at
       ? admin.firestore.Timestamp.fromMillis(subscription.ended_at * 1000)
+      : null,
+    trial_start: subscription.trial_start
+      ? admin.firestore.Timestamp.fromMillis(subscription.trial_start * 1000)
+      : null,
+    trial_end: subscription.trial_end
+      ? admin.firestore.Timestamp.fromMillis(subscription.trial_end * 1000)
       : null,
   };
   await subsDbRef.set(subscriptionData);