+++ firestore-stripe-subscriptions/CHANGELOG.md
@@ -1,3 +1,62 @@
+## Version 0.1.3
+
+[fix] - Apply trial days from pricing plan to the checkout session.
+
+[feat] - Support setting metadata on the subscription:
+
+You can optionally set a metadata object with key-value pairs when creating the checkout session. This can be useful for storing additional information about the customer's subscription. This metadata will be synced to both the Stripe subscription object (making it searchable in the Stripe Dashboard) and the subscription document in the Cloud Firestore.
+
+```js
+const docRef = await db
+  .collection("customers")
+  .doc(currentUser)
+  .collection("checkout_sessions")
+  .add({
+    price: "price_1GqIC8HYgolSBA35zoTTN2Zl",
+    success_url: window.location.origin,
+    cancel_url: window.location.origin,
+    metadata: {
+      item: "item001",
+    },
+  });
+```
+
+[feat] - Sync additional data from the Stripe subscription object to Cloud Firestore:
+
+```ts
+/**
+ * Start of the current period that the subscription has been invoiced for.
+ */
+current_period_start: FirebaseFirestore.Timestamp;
+/**
+ * End of the current period that the subscription has been invoiced for. At the end of this period, a new invoice will be created.
+ */
+current_period_end: FirebaseFirestore.Timestamp;
+/**
+ * A date in the future at which the subscription will automatically get canceled.
+ */
+cancel_at: FirebaseFirestore.Timestamp | null;
+/**
+ * If the subscription has been canceled, the date of that cancellation. If the subscription was canceled with `cancel_at_period_end`, `canceled_at` will still reflect the date of the initial cancellation request, not the end of the subscription period when the subscription is automatically moved to a canceled state.
+ */
+canceled_at: FirebaseFirestore.Timestamp | null;
+/**
+ * If the subscription has a trial, the beginning of that trial.
+ */
+trial_start: FirebaseFirestore.Timestamp | null;
+/**
+ * If the subscription has a trial, the end of that trial.
+ */
+trial_end: FirebaseFirestore.Timestamp | null;
+/**
+ * Set of key-value pairs that you can attach to an object.
+ * This can be useful for storing additional information about the object in a structured format.
+ */
+metadata: {
+  [name: string]: string;
+};
+```
+
 ## Version 0.1.2
 
 [feat] - Adds the `onUserDeleted` function which is triggered by a user being deleted in Firebase Authentication. Upon user deletion the extension will delete their customer object in Stripe which will immediately cancel all subscriptions for the user.